Add files via upload
For the execution of this algorithm an edgelist.csv dataframe is required, this dataframe must have three columns, where the first two contain the list of links and the third the year of publication of the article or collaboration between the authors. The names of these three columns must be: Source, Target and year. Source and Target must have factors and year must have integers, as data class.